fix: refactor folder structure for validators and also tests
