+++ src/main/java/com/plazoleta/foodcourtmicroservice/domain/validation/restaurant/impl/RestaurantNameValidator.java
@@ -1,9 +1,10 @@
-package com.plazoleta.foodcourtmicroservice.domain.validation;
+package com.plazoleta.foodcourtmicroservice.domain.validation.restaurant.impl;
 
 
 import com.plazoleta.foodcourtmicroservice.domain.model.RestaurantModel;
 import com.plazoleta.foodcourtmicroservice.domain.exceptions.InvalidElementFormatException;
 import com.plazoleta.foodcourtmicroservice.domain.utils.constants.DomainMessagesConstants;
+import com.plazoleta.foodcourtmicroservice.domain.validation.AbstractValidator;
 
 public class RestaurantNameValidator extends AbstractValidator<RestaurantModel> {
     @Override
