+++ src/test/java/com/plazoleta/foodcourtmicroservice/domain/validation/restaurant/RestaurantValidatorChainTest.java
@@ -1,9 +1,11 @@
-package com.plazoleta.foodcourtmicroservice.domain.validation;
+package com.plazoleta.foodcourtmicroservice.domain.validation.restaurant;
 
-import com.plazoleta.foodcourtmicroservice.domain.model.RestaurantModel;
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import com.plazoleta.foodcourtmicroservice.domain.model.RestaurantModel;
 
 class RestaurantValidatorChainTest {
