+++ src/test/java/com/plazoleta/foodcourtmicroservice/domain/validation/restaurant/impl/RestaurantRequiredFieldsValidatorTest.java
@@ -1,10 +1,12 @@
-package com.plazoleta.foodcourtmicroservice.domain.validation;
+package com.plazoleta.foodcourtmicroservice.domain.validation.restaurant.impl;
+
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+import org.junit.jupiter.api.Test;
 
 import com.plazoleta.foodcourtmicroservice.domain.exceptions.RequiredFieldsException;
 import com.plazoleta.foodcourtmicroservice.domain.model.RestaurantModel;
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class RestaurantRequiredFieldsValidatorTest {
     private final RestaurantRequiredFieldsValidator validator = new RestaurantRequiredFieldsValidator();
